// src/app/(artikler)/articleConfig.generated.ts
export const articleConfig: PageInfoConfig = {
    arbeidsgivertjenester: {
        title: "Hvem kan bruke arbeidsgivertjenestene?",
        language: "nb",
        proofread: true,
        category: "employer-guides",
        description:
            "Få oversikt over arbeidsgivertjenestene på arbeidsplassen.no og hvordan du kan bruke dem i rekrutteringen.",
        updatedAt: "2025-09-15",
    },
// ...
};